refactor token instance page title
